+++ R/eng-mod_ind_ita.R
@@ -48,7 +48,36 @@ eng_mod_ind_ita_ui <- function(id){
       box(plotlyOutput(ns("tit")),
         width = 4, title = "ICU"
       )
+    ),
+
+    h2(HTML("Intensive care")),
+    fluidRow(
+      selectInput(width = "45%", ns("whichRegion"),
+                  label = "Select regions",
+                  choices = regions(),
+                  selectize = TRUE,
+                  selected = c("Lombardia", "Veneto", "Emilia Romagna"),
+                  multiple = TRUE
+      ),
+      sliderInput(width = "45%", ns("lastDate"),
+                  label = "Select the last day that you want to take into consideration for the estimation of the model",
+                  value = max(dpc_covid19_ita_regioni$data),
+                  min = min(dpc_covid19_ita_regioni$data) + lubridate::days(6),
+                  max = max(dpc_covid19_ita_regioni$data),
+                  step = lubridate::days(1),
+                  animate = animationOptions(interval = 1100)
+      )
+    ),
+    box(plotlyOutput(ns("titamponi")),
+        width = 12,
+        title = "Regional percentage (fit loess, span = 1.5, degree = 2) of population that, repectively, has been admitted to the ICU (vertical axis) and has not been (as of today) hospitalized though being tested (this data was approximated by the total number of tests minus hospitalized patients), up to the day chosen in the slider. It is possible to see the entire evolution automatically by clicking play.",
+        footer = "REMARKS: The curves show the daily cumulative events on the horizontal axis (hence it is never possible to 'go back in time') and daily net on the vertical axis (hence we can both 'move up' and 'move down'). It is important to notice that the curves are 'temporally long': they all have the same number of days (i.e., from the 24th of February, up to the date chosen in the slider) and, moving along the curves, each point represents the data that was observed on the corresponing day."
+    ),
+    box(DT::DTOutput(ns("dt_tamponi")),
+        width = 12,
+        title = "Table: regional evolution of bed occupancy with respect to the asympomatic tests (weighted on the population)"
     )
+
 
   )
 }
@@ -112,6 +141,22 @@ eng_mod_ind_ita_server <- function(id) {
 
   }
 
+  ## dati per tamponi
+  dati_tamponi <- dpc_covid19_ita_regioni %>%
+    dplyr::left_join(region_population) %>%
+    dplyr::group_by(.data$denominazione_regione) %>%
+    dplyr::mutate(
+      tamponi_no_sintomi = .data$tamponi -
+        .data$ricoverati_con_sintomi -
+        .data$terapia_intensiva,
+      tamp_asint_pesati = 100 * (
+        .data$tamponi_no_sintomi/.data$residenti
+      ),
+      intensiva_pesati  = 100 * (
+        .data$terapia_intensiva/.data$residenti
+      )
+    ) %>%
+    dplyr::ungroup()
 
 
   callModule(id = id, function(input, output, session){
@@ -154,6 +199,53 @@ eng_mod_ind_ita_server <- function(id) {
       ggplotly(gg_ind_plot("tit") +
         ylab("% (with respect to the previous day)")
       )
+    })
+
+
+    ## terapia intensiva per tamponi
+    data_to_use <- reactive({
+      req(input$whichRegion)
+      req(input$lastDate)
+
+      dati_tamponi %>%
+        dplyr::filter(
+          .data$data <= input$lastDate,
+          .data$denominazione_regione %in% input$whichRegion
+        )
+    })
+
+
+
+    output$titamponi <- renderPlotly({
+
+      gg_titamponi <- data_to_use() %>%
+        ggplot(aes(
+          x = .data$tamp_asint_pesati,
+          y = .data$intensiva_pesati,
+          colour = .data$denominazione_regione,
+          label = .data$denominazione_regione
+        )) +
+        geom_smooth(method = stats::loess, span = 1.5, se = FALSE) +
+        ylab("(ICU / residents) x 100000 ppl") +
+        xlab("(tests on asympotomatics / residents) x 100000 ppl") +
+        coord_cartesian(
+          xlim = c(0, max(dati_tamponi$tamp_asint_pesati) + 1L),
+          ylim = c(0, max(dati_tamponi$intensiva_pesati))
+        ) +
+        scale_color_discrete(name = "Regione")
+
+      ggplotly(gg_titamponi)
+    })
+
+    output$dt_tamponi <- DT::renderDT({
+      data_to_use() %>%
+        dplyr::select(
+          .data$data, .data$denominazione_regione,
+          .data$tamp_asint_pesati,
+          .data$intensiva_pesati
+        ) %>%
+        dplyr::mutate(data = as.Date(.data$data)) %>%
+        dplyr::mutate_if(is.numeric, round, digits = 2)
     })
 
   })