+++ NEWS.md
@@ -1,3 +1,6 @@
+* focus veneto total hosp 20200327
+* update plot tamponi/intensive (#29)
+
 # covid19ita 0.22.0
 
 * data update