+++ NEWS.md
@@ -1,3 +1,5 @@
+* translation focus veneto hosp 20200328
+*update plot tamponi/intensive eng (#29)
 * focus veneto total hosp 20200327
 * update plot tamponi/intensive (#29)